modularized website logo
